refactored passphrase to use a simpler goodie.txt
Instead of building a hash from a source file
containing index and word separated by spaces
this new version builds an array of words and 
randomly selects elements from it.

This will make it easier to modify the list of 
words since indexes don't need to be maintained
by hand.

This has the added bonus of being able to splice
from the array, thus guaranteeing that we don't
pick the same word more than once in a single
passphrase.