Clean up bold in Average.
I think bold should be generally reserved for matching words, because
otherwise it is confusing with the organic links.

However, I can see a case for bolding the actual answer, and so tried that.

Also moved | to ;, which makes it easier to read.

We could move to three lines and do the Mean: to <i>Mean:</i> like we've
been doing with other plugins.