Jira Goodie first (incomplete) version
When a JIRA bug ticket number is searched, this goodie
checked if the ticket is owned by a project hosted on the Apache JIRA bugtracker
or on the Codehause JIRA bugtracker.
A JIRA ticket is formed by a KEY (corresponding to a project) and
a ticket number. The validity of the key is checked before displaying
any result (comparison with hard-coded keys).

This version is runnable but is still incomplete because not every project key
has been added yet.