Travis uses the Locale::Codes package 3.16 on the testrun with Perl 5.14. This version of the Locale::Codes package does not support the 'eng' language codeword (and probably a lot others) which is used in the iso639 tests.
From version 3.20 onward the Locale::Codes supports this language code. The Travis test with Perl 5.16 is using version 3.21 of Locale::Codes, so it is not affected.

I changed the test file so that it checks the version of Locale::Codes and leaves out the 'iso 639 eng' test for versions lower than 3.20.