Passphrase: replace word list with more common words.
Fixes issue #28

Rather than cleaning up the old list, I decided to replace it wholesale
with a list of more common words in American English.

This list is sourced from the free 5,000 most common word list on
http://www.wordfrequency.info.

The provided list was edited for a normalized format and to eliminate
duplicates and words of under 3 letters.