Calculator: improve scientific notation support.
Fixes #270.

Scientific notation was not supported in all places.  This should fix
that. I admit the regex is a bit much for me to hold in my head, so I am
not sure if there are now incidentally misparsed queries.

All tests pass, including new ones to cover the broken cases mentioned
in the issue.