Binary: tighten up the binary to decimal trigger.
- Do not match when there are non-binary [01] characters in something
  which we posit to be "binary."
- Add test cases for the regressions.

We don't want to match on things which do not appear to be asking for
conversions. In the binary -> decimal case this is somewhat more
ambiguous, except that we can demand that the input look like binary, so
it's actually not so bad.

I didn't pay enough attention to the previous regex and how it matched
up with the new code. :(