Iterate toward more formal query parsing.
This is mostly an attempt to make it easier (for me) to read and
identify everything which needs to be covered in any further
refactoring.

There is one funcational difference:  'c' is no longer an acceptable
constant. I was vaguely bothered that it was the only one with units,
anyway and once it made the parsing of 'cos' turn ugly, I just ripped
it out.

Also added a test for an identifiable regression from a former code
comment.