Calculator: use Math::Trig for trig functions.
Also, to get the value of pi.

In service of this change, also add some additional available functions.
Not so much because I think people were clamoring for secants, but as a
way of expanding expectations for the nascent parser: fix replacement
issues with 'e', better description of what a "function" might look
like, etc.