PublicDNS: put items in a table.
Possibly the most roundabout way to resolve issue #375.

- Remove static files from share directory.
- Put info into a perl structure.
- Generate the two static texts from the structure.
- Add minimal test for proper form.

It's almost like I haven't heard of Text::Table or Template::Toolkit.
Nevertheless, I think this is not too arcane for this relatively simple
usage.  On the other hand, I am going to feel like a fool for not just
editing the two static files when it turns out there are never any
changes to the data.