dist.ini: Add YAML to Prereqs.
Even though it is implicitly required (at present) for duckpan, being
explicit makes it clear that we are actually using it in local code.
Still, we'd rather not make it depend on a particular version, so as
not to conflict (or bloat.)  YAML is pretty stable, so this should
be fine.

Also remove a no-longer-required `use YAML` from the ZappBrannigan
Goodie.