FlipText: Add "text" to the trigger word.
"flip", "mirror", "spin", and "rotate" are all very common words, and this instant answer comes up often for queries such as "flip flops". While we don't have a way of setting individual goodies with low priority (all goodies have high priority at the moment), it's important that we control the query space of each one. So in this case, adding a "text" as a requirement will prevent it from triggering on "flip flop", "back flip", etc.