Calculator: style up the answer a bit.
This mostly just adds style to the HTML answers to be more in line with
other IAs.

The test changes are because, in general, the HTML answer is exactly the
same as the text answer, just wrapped up in some HTML. We're mostly
concerned with if we get correct answers, not how they happen to be
wrapped for HTML display.

In cases where that's different (the <sup> regression) we do more
thorough testing on the contents.

Also, removed one repeated test.

I'm starting to think this "apply_css" and maybe even "wrap_html" should
be in a role somewhere.