Calculator: add 'squared' and 'score'.
This is an attempt to add the requested "squared" operation without all
of the nasty trigger changes I introduced last time I tried.

Also added a constant for 'score' as both an attempt to better
understand that trigger regex and an historical oddity.

It's clear that if I want to continue with this I should make a lexer.