NumberStyle: minimal changes for Conversions.
I am not claiming this is entirely the correct way to do this in this
package, just that it works fairly easily to expand the scope of the
Conversions goodie, as well.

Gonna need feedback from @moollaza and @mintsoft.

On the upside, I found a bug in the styler code by writing a test here!