DatesRole: clarify how we determine the flip.
I think this is vaguely easier to read (and it may return a bit earlier,
if that's important which it is probably not).  I may, of course, be
wrong.

Add named captures for the ambiguous formats to make it easier to know
what we "expect" the order to be unless it is flipped.

Also includes some minor formatting changes, due to my incessant use of
`perltidy`, and some test case arrangement changes to satisfy my desire
for order.