CalendarToday: protect on missing timezone.
DateTime doesn't like to have timezones passed in which it can't parse.
So, just continue on with whatever we have by default if the call
doesn't work.

Remove parsePsuedoDate which is now basically just an alias for
parse_vague_string_todate.

Also, reuse the timezoned DateTime for our default, rather than create
a new object.