UnixTime: update output to include the epoch.
This is in service to some submarine commits which will be arriving soon
to deal with Goodies templating.  This makes UnixTime look "more like it
will" in addition to supplying what we determined to be the supplied
epoch.

This is most useful in the "empty" epoch case where we will tell them
the current epoch which may be of interest.