templates: update to include basic ZCI attributes.
There will likely be some debate about whether or not to include the
`is_cached` value, but I like making it explicit even when there is
a default.

On the `answer_type`, the tests as produced by `duckpan new` could
not pass. It might not be entirely clear to the new developer how to
add the `answer_type` and fix them.