HelpLine.t: do not construct a test plan.
We're not actually getting a test count from test_zci. Someday soon this
will reconcile exactly with the number of `test_zci` calls you make, but
it clearly does not today (see the extra 2 factor).

Rather than try to predict which version is running these tests, remove
the plan for now.