NumberStyle: do not automatically convert numbers with_html.
In some cases, we have numbers which are _already_ in the appropriate
style. Most notably, when considering the user input values from which we
made the style determinations in the first place.

This fixes the bug, but is still unsatisfying because we would need
to convert both in and out in order to do better input styling.