Changed image size, changed font, fixed deps
Changing font and image size required customized script, but that means
it's mostly self contained now.

The .ttf version of Proxima Nova has to be in the share directory so the
generate script can use it (I wasn't sure if/where Proxima Nova is on
the server).

Added libgd-dev to dist.ini.