Improve triggers
Vastly improves the triggering!

I had a look at searches related to significant figures and none of the
common searches involved `sd` or `sig digs`, so I got rid of those
triggers.

A lot of the searches were along the lines of "What are the significant
figures of X?" - so I've updated the handler to allow for these types of
queries.