Remove fallback to 'singular_property' for plurals
Now 'prefer_opt' can take arrays of the form [option, sub] where sub
takes the value of 'option' to produce a result. This also means we can
know what the option name is (for error messages), whereas before we had
to remove the CODE references from the arguments.

This should hopefully work better than before as it won't *force* a
plural version of the property - only if 'plural_property' or 'property'
is specified will a plural match be allowed.