Add support for 'ignore'd phrases
These allow categories (and/or cheat sheets) to specify phrases that
should be ignored when triggering.

For example, many language cheat sheets use 'english to X' as an alias,
but this would be better placed in 'ignore'. This would allow these
cheat sheets to trigger on 'english to X cheat sheet', 'english to X
phrases', 'X phrases english to' (you decide whether you like that last
one ;)) etc...

In essence, the 'ignore' attribute defines an array of phrases that
should be entirely stripped from the query when that cheat sheet or
category triggers.