Fix issue with very large inputs to trig functions
As pointed out by @mintsoft, queries such as `sin(10,000,000,000)` were
incorrectly returning 0; this was due to the way Math::Cephes was
handling the numbers internally.

We fix this by modding the input by 2 * Pi (a.k.a, Tau - long live Tau);
which brings it back into a range that can be handled by Math::Cephes.

I have chosen not to do this for every input as it brings in more
operations that for the most part, are not required.